+++ src/DisallowListerException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Accentinteractive;
+
+class DisallowListerException extends \Exception {
+
+}